requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.4
django-autoslug==1.9.9
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-js-asset==2.1.0
Pillow==10.0.0
sqlparse==0.4.4
tzdata==2023.3
